pyproject.toml
[project]
name = "testOrness"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)"
]

[tool.poertry]
packages = [{include = "testOrness"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
